some fx

+++ tensor/apex_tensor_cpu.cpp
@@ -196,7 +196,7 @@ namespace apex_tensor{
         
         void shuffle( std::vector<CTensor3D> &data ){
             for( size_t i = 0 ; i < data.size() ; i ++ ){
-                int j = (int)apex_random::next_uint32( data.size() );
+                int j = (int)apex_random::next_uint32( (uint32_t)data.size() );
                 CTensor3D a = data[i];
                 data[i] = data[j]; 
                 data[j] = a;
